test(core): migrate SnapshotTests (Wave B)
32 tests, 73 failure sites. Pins the AppInitFile / snapshot wire
format: decode/validate, version gating (v2 only), pane-id and tab-id
uniqueness checks, normalized selectedTabId, optional-id minting,
surface launch resolution, scrollback backward-compat, tab color
round-trip, preferences-draft dropping, and the todo round-trips on
both panes and tabs.

Gate results:
- NAME parity:      32/32 baseline names found in migrated @Tests
- ASSERTION parity: baseline TOTAL 73 == new TOTAL 73
                    (69 #expect + 4 Issue.record). Each legacy
                    do/catch `throw TestFailure` for the
                    "expected throw but got success" arm becomes
                    `Issue.record + return`, preserving the 2 sites
                    per error-expecting test (1 throw + 1 expectEqual
                    -> 1 Issue.record + 1 #expect). The typed-catch
                    arm propagates unmatched error types as the test
                    function's thrown error, mirroring the legacy
                    behavior.
- STYLE:            32/32 tests carry Intent/Why/Scenario preambles

Verified locally:
- swift test --package-path lib/DanTermCore -> 105/105 in 9 suites (0.61s)
- ./test.sh -> 973/973 (= 1005 - 32)
